+++ toothpick-runtime/src/test/java/toothpick/concurrency/ToothpickManipulationsMultiThreadTest2.java
@@ -1,4 +1,4 @@
-package toothpick;
+package toothpick.concurrency;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +7,8 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import toothpick.Toothpick;
+import toothpick.ToothpickVisibilityExposer;
 import toothpick.concurrency.threads.AddSameScopeThread;
 import toothpick.concurrency.threads.AddScopeToListThread;
 import toothpick.concurrency.threads.RemoveSameScopeThread;
@@ -20,7 +22,7 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static toothpick.concurrency.utils.ThreadTestUtil.STANDARD_THREAD_COUNT;
 
-public class ToothpickManipulationsMultiThreadTest {
+public class ToothpickManipulationsMultiThreadTest2 {
 
   static final String ROOT_SCOPE = "ROOT_SCOPE";
   final List<Object> scopeNames = new CopyOnWriteArrayList<>();
@@ -130,7 +132,7 @@ public class ToothpickManipulationsMultiThreadTest {
     for (TestableThread thread : threadList) {
       assertTrue(String.format("test of thread %s failed", thread.getName()), thread.isSuccessful());
     }
-    assertThat(Toothpick.getScopeNamesSize(), is(2));
+    assertThat(ToothpickVisibilityExposer.getScopeNamesSize(), is(2));
   }
 
   @Test
@@ -158,6 +160,6 @@ public class ToothpickManipulationsMultiThreadTest {
     for (TestableThread thread : threadList) {
       assertTrue(String.format("test of thread %s failed", thread.getName()), thread.isSuccessful());
     }
-    assertThat(Toothpick.getScopeNamesSize(), anyOf(is(1), is(2)));
+    assertThat(ToothpickVisibilityExposer.getScopeNamesSize(), anyOf(is(1), is(2)));
   }
 }
